refactor: make all internal modules private with underscore prefix
Prevents users from importing internal modules (git, hunk, lines, patch,
ui, cli) as public API, allowing us to freely change internals without
breaking downstream code. Only __version__ remains public via __init__.py.
