+++ git_hunk/__main__.py
@@ -1,3 +1,3 @@
-from git_hunk.cli import cli
+from git_hunk._cli import cli
 
 cli()
